Move 'Hive Metadata Scan: Support case insensitive name mapping' (PR 52) to hivelink-core (#102)
