+++ core/src/main/java/org/apache/iceberg/TableMetadata.java
@@ -130,18 +130,6 @@ public class TableMetadata implements Serializable {
         .build();
   }
 
-  public static TableMetadata newTableMetadataWithoutFreshIds(Schema schema,
-      PartitionSpec spec,
-      String location,
-      Map<String, String> properties) {
-    return new TableMetadata(null, DEFAULT_TABLE_FORMAT_VERSION, UUID.randomUUID().toString(), location,
-        INITIAL_SEQUENCE_NUMBER, System.currentTimeMillis(),
-        -1, schema, INITIAL_SPEC_ID, ImmutableList.of(spec),
-        SortOrder.unsorted().orderId(), ImmutableList.of(SortOrder.unsorted()),
-        ImmutableMap.copyOf(properties), -1, ImmutableList.of(),
-        ImmutableList.of(), ImmutableList.of());
-  }
-
   public static class SnapshotLogEntry implements HistoryEntry {
     private final long timestampMillis;
     private final long snapshotId;
@@ -257,7 +245,7 @@ public class TableMetadata implements Serializable {
   private final List<MetadataUpdate> changes;
 
   @SuppressWarnings("checkstyle:CyclomaticComplexity")
-  TableMetadata(String metadataFileLocation,
+  TableMetadata(InputFile file,
                 int formatVersion,
                 String uuid,
                 String location,
