+++ hivelink-core/src/main/java/org/apache/iceberg/hivelink/core/schema/MergeHiveSchemaWithAvro.java
@@ -44,7 +44,7 @@ import org.apache.iceberg.relocated.com.google.common.base.Preconditions;
  * 2. Copies field names, nullability, default value, field props from the Avro schema
  * 3. Copies field type from the Hive schema.
  *    TODO: We should also handle some cases of type promotion where the types in Avro are potentially more correct
- *    e.g.BINARY in Hive -> FIXED in Avro, STRING in Hive -> ENUM in Avro, etc
+ *    e.g.BINARY in Hive to FIXED in Avro, STRING in Hive to ENUM in Avro, etc
  * 4. Retains fields found only in the Hive schema; Ignores fields found only in the Avro schema
  * 5. Fields found only in Hive schema are represented as optional fields in the resultant Avro schema
  * 6. For fields found only in Hive schema, field names are sanitized to make them compatible with Avro identifier spec
